+++ HISTORY.rst
@@ -4,6 +4,8 @@ History
 
 Unreleased
 ------------------
+* Added dots enabling in config.cfg
+* Added coverage
 * Added config as a singleton.
 * Added conf path passing from console argument.
 * Added testing conf and application fixtures.